fixed problem where only one or two members shown
when the member thumbnails are about to loop, only one or two used to be shown.  Now 4 will always be shown.  It's somewhat hackish, because any members are chosen to fill in the gaps, so there can be repeat members in adjacent transitions. (ie, Ken might show up in 3 consecutive slides), but it works and it looks nice.